LayoutTests/fast/text/font-face-family.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
// The family attribute should return the serialization of the parsed value.
shouldBeEqualToString("(new FontFace('a', 'url(garbage.otf)')).family", "a");
shouldBeEqualToString("(new FontFace('4a', 'url(garbage.otf)')).family", "\"4a\"");
shouldBeEqualToString("(new FontFace('4\"a', 'url(garbage.otf)')).family", "\"4\\\"a\"");
shouldBeEqualToString("(new FontFace('4\\'a', 'url(garbage.otf)')).family", "\"4'a\"");
shouldBeEqualToString("(new FontFace('4\\'a\"b', 'url(garbage.otf)')).family", "\"4'a\\\"b\"");
shouldBeEqualToString("(new FontFace('ab', 'url(garbage.otf)')).family", "ab");
shouldBeEqualToString("(new FontFace('\"ab\"', 'url(garbage.otf)')).family", "\"\\\"ab\\\"\"");
shouldBeEqualToString("(new FontFace('a b', 'url(garbage.otf)')).family", "a b");
shouldBeEqualToString("(new FontFace('a b, c', 'url(garbage.otf)')).family", "\"a b, c\"");
shouldBeEqualToString("(new FontFace('ab,c', 'url(garbage.otf)')).family", "\"ab,c\"");
</script>
</body>
</html>
