LayoutTests/fast/text/font-face-family-expected.txt

PASS (new FontFace('a', 'url(garbage.otf)')).family is "a"
PASS (new FontFace('4a', 'url(garbage.otf)')).family is "\"4a\""
PASS (new FontFace('4"a', 'url(garbage.otf)')).family is "\"4\\\"a\""
PASS (new FontFace('4\'a', 'url(garbage.otf)')).family is "\"4'a\""
PASS (new FontFace('4\'a"b', 'url(garbage.otf)')).family is "\"4'a\\\"b\""
PASS (new FontFace('ab', 'url(garbage.otf)')).family is "ab"
PASS (new FontFace('"ab"', 'url(garbage.otf)')).family is "\"\\\"ab\\\"\""
PASS (new FontFace('a b', 'url(garbage.otf)')).family is "a b"
PASS (new FontFace('a b, c', 'url(garbage.otf)')).family is "\"a b, c\""
PASS (new FontFace('ab,c', 'url(garbage.otf)')).family is "\"ab,c\""
PASS successfullyParsed is true

TEST COMPLETE
